Move bytes-lit literal migration note to v26->v27
This change releases with v27, so document it under the
'Migrating from v26 to v27' section.
