+++ soroban-sdk/src/_migrating/v27_bytes_literals.rs
